Refactor OBV into class
