+++ speculator/tests/unit/test_obv.py
@@ -1,4 +1,4 @@
-from speculator.features import obv
+from speculator.features.OBV import OBV
 import unittest
 
 class OBVTest(unittest.TestCase):
@@ -6,17 +6,16 @@ class OBVTest(unittest.TestCase):
         # Case 1: close > close_prev, return obv + volume
         curr = {'close': 5, 'volume': 1}
         prev = {'close': 4, 'obv': 2}
-        self.assertEqual(obv.eval_algorithm(curr, prev), 3)
-        
+        self.assertEqual(OBV.eval_algorithm(curr, prev), 3)
+
     def test_eval_algorithm_case2(self):
         # Case 2: close == close_prev, return obv
         curr = {'close': 5, 'volume': 1}
         prev = {'close': 5, 'obv': 2}
-        self.assertEqual(obv.eval_algorithm(curr, prev), 2)
+        self.assertEqual(OBV.eval_algorithm(curr, prev), 2)
 
     def test_eval_algorithm_case3(self):
         # Case 3: close < close_prev, return obv - volume
         curr = {'close': 5, 'volume': 1}
         prev = {'close': 6, 'obv': 2}
-        self.assertEqual(obv.eval_algorithm(curr, prev), 1)
-
+        self.assertEqual(OBV.eval_algorithm(curr, prev), 1)
